requirements.txt
arabic-reshaper==2.1.3
asgiref==3.4.1
Brotli==1.0.9
cffi==1.14.6
cssselect2==0.4.1
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.2.6
django-import-export==2.6.1
django-translations==1.3.0
djangorestframework==3.12.4
et-xmlfile==1.1.0
fonttools==4.26.2
future==0.18.2
html5lib==1.1
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.9
Pillow==8.3.2
pycparser==2.20
pydyf==0.1.1
PyPDF2==1.26.0
pyphen==0.11.0
python-bidi==0.4.2
pytz==2021.1
PyYAML==5.4.1
reportlab==3.6.1
six==1.16.0
sqlparse==0.4.1
tablib==3.0.0
tinycss2==1.1.0
weasyprint==53.1
webencodings==0.5.1
xhtml2pdf==0.2.5
xlrd==2.0.1
xlwt==1.3.0
zopfli==0.1.8
